Best score =  3.754. (Sharpe, Calmar) = (1.986, 1.768).
  retMean retStd     hpr retAnnual volAnnual sharpe calmar  score    mdd      mddT       ret  id
0   0.034  0.275  103.48      8.62      4.34  1.986  1.768  3.754  4.877  20220928  Opn005L1  s0
2   0.032  0.273   92.53      7.96      4.32  1.843  1.581  3.424  5.032  20200203       NaN  p0
1   0.029  0.280   82.08      7.29      4.42  1.648  1.198  2.846  6.088  20200203  Cls005L1  s1

+ unify the rate in vt_test and signals_facotrs = 1.00

+++ run_factor.ps1
@@ -4,15 +4,18 @@ param (
     [string]$bgn_date_qtest,
     [string]$stp_date,
     [string]$factor,
-    [Switch]$DisableMP
+    [Switch]$DisableMP,
+    [Switch]$TestOT
 )
 
-function CheckExistenceRemove {
+function CheckExistenceRemove
+{
     param (
         [string]$TargetPath
     )
 
-    if (Test-Path $TargetPath) {
+    if (Test-Path $TargetPath)
+    {
         Remove-Item $TargetPath -Recurse
     }
 }
@@ -33,21 +36,30 @@ Remove-Item "$proj_dir\vt_tests\data\$factor-*.db"
 Remove-Item "$proj_dir\vt_tests\plots\$factor-*.pdf"
 Remove-Item "$proj_dir\vt_tests\\$factor-*.csv"
 
-Remove-Item "$proj_dir\ot_tests\data\$factor-*.db"
-Remove-Item "$proj_dir\ot_tests\plots\$factor-*.pdf"
-Remove-Item "$proj_dir\ot_tests\\$factor-*.csv"
+if ($TestOT)
+{
+    Remove-Item "$proj_dir\ot_tests\data\$factor-*.db"
+    Remove-Item "$proj_dir\ot_tests\plots\$factor-*.pdf"
+    Remove-Item "$proj_dir\ot_tests\\$factor-*.csv"
+}
 
 if ($DisableMP)
 {
     python main.py --bgn $bgn_date_factor --stp $stp_date --nomp factor --fclass $factor
     python main.py --bgn $bgn_date_qtest --stp $stp_date --nomp ic --fclass $factor
     python main.py --bgn $bgn_date_qtest --stp $stp_date --nomp vt --fclass $factor
-    python main.py --bgn $bgn_date_qtest --stp $stp_date --nomp ot --fclass $factor
+    if ($TestOT)
+    {
+        python main.py --bgn $bgn_date_qtest --stp $stp_date --nomp ot --fclass $factor
+    }
 }
 else
 {
     python main.py --bgn $bgn_date_factor --stp $stp_date factor --fclass $factor
     python main.py --bgn $bgn_date_qtest --stp $stp_date ic --fclass $factor
     python main.py --bgn $bgn_date_qtest --stp $stp_date vt --fclass $factor
-    python main.py --bgn $bgn_date_qtest --stp $stp_date ot --fclass $factor
+    if ($TestOT)
+    {
+        python main.py --bgn $bgn_date_qtest --stp $stp_date ot --fclass $factor
+    }
 }
